+++ projects/ui/src/lib/components/otp-input/index.ts
@@ -1,0 +1,1 @@
+export * from './kui-otp-input.component';